Updated the proof behavior for unit clauses
Now, when a unit clause is directly imported, it adds the
clause ID to the proof chain *after* adding the unit with empty
chain to the proof.

This means that any unit clauses that are derived from the
unit correctly reference the unit, which was not true in all
cases previously.